+++ CHANGELOG.md
@@ -88,6 +88,8 @@
 * Ctrl+wheel up/down bound to `font-increase` and `font-decrease`
   respectively (in addition to the already existing default key
   bindings `ctrl-+` and `ctrl+-`).
+* Use XRGB pixel format (instead of ARGB) when there is no
+  transparency.
 
 [1526]: https://codeberg.org/dnkl/foot/issues/1526
 [1528]: https://codeberg.org/dnkl/foot/issues/1528